requirements.txt
tornado
pymongo
motor
dictshield
requests
PyJWT
python-memcached
